Fixed incorrect datatype for a value in BBLRecord;  Added BBL prepared statement for insert; Added utility methods for creating and extracting string arrays

+++ src/net/skyebook/padloader/read/BBLReader.java
@@ -99,7 +99,7 @@ public class BBLReader extends CSVReader {
 				bbl.setHilot(Integer.parseInt(values[i]));
 				break;
 			case hibblscc:
-				bbl.setHibblscc(Integer.parseInt(values[i]));
+				bbl.setHibblscc(Short.parseShort(values[i]));
 				break;
 			case boro:
 				bbl.setBoro(Short.parseShort(values[i]));
